+++ src/gql/arguments/elements/EpisodeArguments.php
@@ -4,6 +4,7 @@ namespace vnali\studio\gql\arguments\elements;
 
 use Craft;
 use craft\gql\base\ElementArguments;
+use GraphQL\Type\Definition\Type;
 use vnali\studio\elements\Episode;
 use vnali\studio\Studio;
 
@@ -13,7 +14,38 @@ class EpisodeArguments extends ElementArguments
     // =========================================================================
     public static function getArguments(): array
     {
-        return array_merge(parent::getArguments(), self::getContentArguments(), []);
+        return array_merge(parent::getArguments(), self::getContentArguments(), [
+            'episodeSeason' => [
+                'name' => 'episodeSeason',
+                'type' => Type::INT(),
+                'description' => 'Episode Season',
+            ],
+            'episodeType' => [
+                'name' => 'episodeType',
+                'type' => Type::STRING(),
+                'description' => 'Episode type',
+            ],
+            'blocked' => [
+                'name' => 'blocked',
+                'type' => Type::boolean(),
+                'description' => 'Episode block',
+            ],
+            'explicit' => [
+                'name' => 'explicit',
+                'type' => Type::boolean(),
+                'description' => 'Episode Explicit',
+            ],
+            'rss' => [
+                'name' => 'rss',
+                'type' => Type::boolean(),
+                'description' => 'Publish on RSS',
+            ],
+            'podcastId' => [
+                'name' => 'podcastId',
+                'type' => Type::INT(),
+                'description' => 'Podcast ID',
+            ],
+        ]);
     }
 
     /**